Add the possibility to check links
This commit introduces the possibility to check links with `zola check`
by adding two environment variables: `CHECK_LINKS` and `CHECK_FLAGS`.

Similarly to `BUILD_ONLY` and `BUILD_FLAGS`, those two environment
variables allow to trigger the use of `zola check` when running the
GitHub Action and specify flags to be passed to the command, like
`--drafts`.